Create FlyingSaucer.gd
Added the script. Hopefully this works this time. I commented a couple things out as I am not sure if we will want to do this stuff yet. So if we want to add sound effects or such, we just have to add the correct file and uncomment some lines